Move WebTools calling code to a shared layer
