+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting_NetFx/FormattingLanguageServerClient.cs
@@ -3,25 +3,17 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.AspNetCore.Razor.LanguageServer.Hosting;
-using Microsoft.AspNetCore.Razor.PooledObjects;
-using Microsoft.AspNetCore.Razor.Test.Common.Mef;
 using Microsoft.AspNetCore.Razor.Utilities;
-using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Razor;
+using Microsoft.CodeAnalysis.Razor.Formatting;
 using Microsoft.CodeAnalysis.Razor.Logging;
 using Microsoft.CodeAnalysis.Razor.Protocol;
 using Microsoft.CodeAnalysis.Razor.Protocol.Formatting;
-using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
-using Microsoft.VisualStudio.Text;
-using Microsoft.VisualStudio.Utilities;
-using Microsoft.WebTools.Languages.Shared.ContentTypes;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
@@ -40,52 +32,28 @@ internal class FormattingLanguageServerClient(ILoggerFactory loggerFactory) : IC
         _documents.Add("/" + path, codeDocument);
     }
 
-    private Task<RazorDocumentFormattingResponse> FormatAsync(DocumentOnTypeFormattingParams @params)
+    private async Task<RazorDocumentFormattingResponse> FormatAsync(DocumentOnTypeFormattingParams @params)
     {
         var generatedHtml = GetGeneratedHtml(@params.TextDocument.Uri);
-        var generatedHtmlSource = SourceText.From(generatedHtml, Encoding.UTF8);
-        var absoluteIndex = generatedHtmlSource.GetRequiredAbsoluteIndex(@params.Position);
 
-        var request = $$"""
-            {
-                "Options":
-                {
-                    "UseSpaces": {{(@params.Options.InsertSpaces ? "true" : "false")}},
-                    "TabSize": {{@params.Options.TabSize}},
-                    "IndentSize": {{@params.Options.TabSize}}
-                },
-                "Uri": "{{@params.TextDocument.Uri}}",
-                "GeneratedChanges": [],
-                "OperationType": "FormatOnType",
-                "SpanToFormat":
-                {
-                    "Start": {{absoluteIndex}},
-                    "End": {{absoluteIndex}}
-                }
-            }
-            """;
+        var edits =  await HtmlFormatting.GetOnTypeFormattingEditsAsync(_loggerFactory, @params.TextDocument.Uri, generatedHtml, @params.Position, @params.Options.InsertSpaces, @params.Options.TabSize);
 
-        return CallWebToolsApplyFormattedEditsHandlerAsync(request, @params.TextDocument.Uri, generatedHtml);
+        return new()
+        {
+            Edits = edits
+        };
     }
 
-    private Task<RazorDocumentFormattingResponse> FormatAsync(DocumentFormattingParams @params)
+    private async Task<RazorDocumentFormattingResponse> FormatAsync(DocumentFormattingParams @params)
     {
         var generatedHtml = GetGeneratedHtml(@params.TextDocument.Uri);
 
-        var request = $$"""
-            {
-                "Options":
-                {
-                    "UseSpaces": {{(@params.Options.InsertSpaces ? "true" : "false")}},
-                    "TabSize": {{@params.Options.TabSize}},
-                    "IndentSize": {{@params.Options.TabSize}}
-                },
-                "Uri": "{{@params.TextDocument.Uri}}",
-                "GeneratedChanges": [],
-            }
-            """;
+        var edits = await HtmlFormatting.GetDocumentFormattingEditsAsync(_loggerFactory, @params.TextDocument.Uri, generatedHtml, @params.Options.InsertSpaces, @params.Options.TabSize);
 
-        return CallWebToolsApplyFormattedEditsHandlerAsync(request, @params.TextDocument.Uri, generatedHtml);
+        return new()
+        {
+            Edits = edits
+        };
     }
 
     private string GetGeneratedHtml(Uri uri)
@@ -93,51 +61,6 @@ internal class FormattingLanguageServerClient(ILoggerFactory loggerFactory) : IC
         var codeDocument = _documents[uri.GetAbsoluteOrUNCPath()];
         var generatedHtml = codeDocument.GetHtmlDocument().GeneratedCode;
         return generatedHtml.Replace("\r", "").Replace("\n", "\r\n");
-    }
-
-    private async Task<RazorDocumentFormattingResponse> CallWebToolsApplyFormattedEditsHandlerAsync(string serializedValue, Uri documentUri, string generatedHtml)
-    {
-        var exportProvider = TestComposition.Editor.ExportProviderFactory.CreateExportProvider();
-        var contentTypeService = exportProvider.GetExportedValue<IContentTypeRegistryService>();
-
-        if (!contentTypeService.ContentTypes.Any(t => t.TypeName == HtmlContentTypeDefinition.HtmlContentType))
-        {
-            contentTypeService.AddContentType(HtmlContentTypeDefinition.HtmlContentType, [StandardContentTypeNames.Text]);
-        }
-
-        var textBufferFactoryService = (ITextBufferFactoryService3)exportProvider.GetExportedValue<ITextBufferFactoryService>();
-        var bufferManager = WebTools.BufferManager.New(contentTypeService, textBufferFactoryService, []);
-        var logger = _loggerFactory.GetOrCreateLogger("ApplyFormattedEditsHandler");
-        var applyFormatEditsHandler = WebTools.ApplyFormatEditsHandler.New(textBufferFactoryService, bufferManager, logger);
-
-        // Make sure the buffer manager knows about the source document
-        var textSnapshot = bufferManager.CreateBuffer(
-            documentUri: documentUri,
-            contentTypeName: HtmlContentTypeDefinition.HtmlContentType,
-            initialContent: generatedHtml,
-            snapshotVersionFromLSP: 0);
-
-        var requestContext = WebTools.RequestContext.New(textSnapshot);
-
-        var request = WebTools.ApplyFormatEditsParam.DeserializeFrom(serializedValue);
-        var response = await applyFormatEditsHandler.HandleRequestAsync(request, requestContext, CancellationToken.None);
-
-        var sourceText = SourceText.From(generatedHtml);
-
-        using var edits = new PooledArrayBuilder<TextEdit>();
-
-        foreach (var textChange in response.TextChanges)
-        {
-            var span = new TextSpan(textChange.Position, textChange.Length);
-            var edit = VsLspFactory.CreateTextEdit(sourceText.GetRange(span), textChange.NewText);
-
-            edits.Add(edit);
-        }
-
-        return new()
-        {
-            Edits = edits.ToArray()
-        };
     }
 
     public async Task<TResponse> SendRequestAsync<TParams, TResponse>(string method, TParams @params, CancellationToken cancellationToken)
